python: rename modules package
It used to be "dagger.server", then we changed to environments and it was renamed to "dagger.env", and finally, before the change to modules, this was renamed more generally to "dagger.ext" (for extensions).

Renaming again to "dagger.mod", and removing the previous concept of multiple types of entrypoints to just "functions".
