+++ sdk/python/requirements.txt
@@ -12,9 +12,9 @@ attrs==23.1.0
     # via cattrs
 backoff==2.2.1
     # via gql
-beartype==0.16.1
+beartype==0.16.4
     # via -r -
-black==23.9.1
+black==23.10.1
     # via -r -
 cattrs==23.1.2
     # via -r -
@@ -59,7 +59,7 @@ packaging==23.2
     #   pytest
 pathspec==0.11.2
     # via black
-platformdirs==3.10.0
+platformdirs==3.11.0
     # via
     #   -r -
     #   black
@@ -67,7 +67,7 @@ pluggy==1.3.0
     # via pytest
 pygments==2.16.1
     # via rich
-pytest==7.4.2
+pytest==7.4.3
     # via
     #   -r -
     #   pytest-httpx
@@ -78,13 +78,13 @@ pytest-httpx==0.26.0
     # via -r -
 pytest-lazy-fixture==0.6.3
     # via -r -
-pytest-mock==3.11.1
+pytest-mock==3.12.0
     # via -r -
 pytest-subprocess==1.5.0
     # via -r -
 rich==13.6.0
     # via -r -
-ruff==0.0.290
+ruff==0.1.2
     # via -r -
 sniffio==1.3.0
     # via
