+++ sdk/python/src/dagger/mod/__init__.py
@@ -1,10 +1,11 @@
 from ._arguments import Argument as Argument
 from ._module import Module as Module
 
-_env = Module()
-function = _env.function
+_default_mod = Module()
+
+function = _default_mod.function
 
 
 def default_module() -> Module:
     """Return the default Module builder instance."""
-    return _env
+    return _default_mod
